A few more tests passing
